third_party/WebKit/LayoutTests/fast/dom/Window/property-access-in-closure-after-navigation-expected.txt

Test that an inner global captured in a closure acts as if its creating context is detached when the creating context is navigated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window is self
PASS window is frames
PASS parent is top
PASS window is null.
PASS self is null.
PASS frames is null.
PASS parent is null.
PASS top is null.
